Finished Security Config
Finished Security Config

+++ backend/src/main/java/com/tesi/francescosasso/finapi/fin_api/api/UserResource.java
@@ -28,7 +28,7 @@ public class UserResource {
     }
 
     @GetMapping("/{id}")
-    public UserDTO getUserById(@PathVariable Long id) {
+    public UserDTO getUserById(@PathVariable Integer id) {
         return userService.findById(id).orElse(null);
     }
 }
