+++ backend/src/main/java/com/tesi/francescosasso/finapi/fin_api/service/UserService.java
@@ -13,5 +13,5 @@ import java.util.Optional;
 public interface UserService {
     List<UserDTO> findAll();
 
-    Optional<UserDTO> findById(Long id);
+    Optional<UserDTO> findById(Integer id);
 }
